Add ByteStream
This change introduces the ByteStream component and associated unit
tests.  It updates the MemoryBuffer component to enable the ByteStream
component to block waiting for changes.

This also introduces a new interface -- OutputBuffer.  This is used to
decouple the ByteStream from the concrete OutputBuffer.  That way, if at
some time in the future we wanted to support writing to something else
(e.g., a file), we could reuse the ByteStream component without
modification.