Add os adaptation layer
Directly making calls to os.XXX APIs is difficult to unit test.  We
wouldn't, for example, want a unit test to be manipulating the
filesystem.

This change adds a simple adaptation layer over the os APIs.  Clients of
the os.* function will use this adaptation layer.  The adaptation layer,
by default, invokes the corresponding os.* function; however, users can
inject a stub function to call instead.  Unit tests can use that to
ensure that the expected calls are made to the os.* functions, in the
expected order, with the expected parameters.